Update multisig-tx.md
Nice. I like the error checking. Definitely makes it feel more polished and useful. 

Interesting issue with the index of address generation. Is there anyway to keep track of that if these are being run in separate environments (so you can't always `createAddress` at the same time? 

Maybe an `assert` around 415 to check the addresses are the same? 

You could also link to the docs on the functions where you talk about adding keys (around line 423-427)

Line 439: What do you mean we don't need this information? Which information?